+++ PlotTalk/accountmanager.h
@@ -11,12 +11,16 @@ class AccountManagerData;
 class AccountManager
 {
 private:
+    static AccountManager* instance;
     AccountManager();
     AccountManager(const AccountManager &);
     AccountManager &operator=(const AccountManager &);
     ~AccountManager();
     QSharedDataPointer<AccountManagerData> data;
 public:
+
+    static *getInstance();
+    bool createAccount(QString &first, QString &last, QString &email, QString &password);
 
 };
